ProcessWire-3.x support only; namespaced module files and demos; can now pass all ProcessWire comments form render options via renderComments() - see CommentForm.php; code cleanup; Updated README; Bumped version to 2.4.2

+++ BlogCleanup.php
@@ -1,4 +1,4 @@
-<?php
+<?php namespace ProcessWire;
 
 /**
 *
@@ -12,7 +12,7 @@
 *	Role
 *
 * @author Francis Otieno (Kongondo)
-* @version 2.4.0
+* @version 2.4.2
 *
 * https://github.com/kongondo/Blog
 * Created February 2014
@@ -39,10 +39,14 @@ class BlogCleanup extends ProcessBlog {
 		$this->removeBlogFiles = $input->remove_tpl_files;
 
 		// was the right button pressed
-		if($cleanupBtn && $cleanupBtn == 'Cleanup') {
+		//if($cleanupBtn && $cleanupBtn == 'Cleanup') {
+		if($cleanupBtn) {
 
 			// Get the module config data
-			$this->data = $this->wire('modules')->getModuleConfigData(get_parent_class($this));
+			#$this->data = $this->wire('modules')->getModuleConfigData(get_parent_class($this));
+			// @note: using strrchr to account for namespaced classes
+			$baseClass = substr(strrchr('\\'.get_parent_class($this), '\\'), 1);
+			$this->data = $this->wire('modules')->getModuleConfigData($baseClass);
 
 			$this->blogStyle = $this->data['blogStyle'];// selected blog style (1-4)
 			$this->commentsUse = $this->data['commentsUse'];// commenting feature on/off
@@ -68,48 +72,46 @@ class BlogCleanup extends ProcessBlog {
 		// grab the main blog parent pages IDs where blogStyle == 3.
 		// Similar for if blogStyle == 4 but we'll unset 'blog-posts' in that case
 		$pagesArray = array(
-
-						$data['blog-posts'],
-						$data['blog-categories'],
-						$data['blog-tags'],
-						$data['blog-comments'],
-						$data['blog-widgets'],
-						$data['blog-authors'],
-						$data['blog-archives'],
-						$data['blog-settings'],
-
+			$data['blog-posts'],
+			$data['blog-categories'],
+			$data['blog-tags'],
+			$data['blog-comments'],
+			$data['blog-widgets'],
+			$data['blog-authors'],
+			$data['blog-archives'],
+			$data['blog-settings'],
 		);
 
 		// if blogStyle == 1 or blogStyle == 2: 'blog' is main parent of all blog pages
 		if ($this->blogStyle == 1 || $this->blogStyle == 2) {
-				$p = $pages->get($data['blog']);
-				// we only proceed if we found the page /blog/
-				// recursively delete the blog page - i.e., including its children
-				if ($p->id)	$pages->delete($p, true);
+			$p = $pages->get($data['blog']);
+			// we only proceed if we found the page /blog/
+			// recursively delete the blog page - i.e., including its children
+			if ($p->id)	$pages->delete($p, true);
 		}
 
 		// if blogStyle == 3: root is the main parent of all blog pages
 		elseif ($this->blogStyle == 3) {
-				foreach ($pagesArray as $pageName) {
-						$p = $pages->get($pageName);
-						// recursively delete the blog pages - i.e., including their children
-						if ($p->id) $pages->delete($p, true);
-				}
+			foreach ($pagesArray as $pageName) {
+				$p = $pages->get($pageName);
+				// recursively delete the blog pages - i.e., including their children
+				if ($p->id) $pages->delete($p, true);
+			}
 		}
 
 		// if blogStyle == 4: root is the main parent of all blog pages but 'blog-post' pages live directly under root
 		elseif ($this->blogStyle == 4) {
-				// remove 'blog-posts' since that page does not exist in this case
-				unset($pagesArray[0]);
-				foreach ($pagesArray as $pageName) {
-							$p = $pages->get($pageName);
-							// recursively delete the blog pages - i.e., including their children
-							if ($p->id) $pages->delete($p, true);
-				}
+			// remove 'blog-posts' since that page does not exist in this case
+			unset($pagesArray[0]);
+			foreach ($pagesArray as $pageName) {
+				$p = $pages->get($pageName);
+				// recursively delete the blog pages - i.e., including their children
+				if ($p->id) $pages->delete($p, true);
+			}
 
-				// additionally in this case since parent of each blog post is root,
-				// we delete all pages using the template 'blog-post'
-				foreach ($pages->find('template=blog-post, include=all') as $p) {if ($p->id) $pages->delete($p);}
+			// additionally in this case since parent of each blog post is root,
+			// we delete all pages using the template 'blog-post'
+			foreach ($pages->find('template=blog-post, include=all') as $p) {if ($p->id) $pages->delete($p);}
 		}
 
 		return $this->cleanUpRepeater();
@@ -123,11 +125,16 @@ class BlogCleanup extends ProcessBlog {
 	*
 	*/
 	private function cleanUpRepeater() {
+
 		$pages = $this->wire('pages');
 		// we delete our repeater page: admin/repeaters/for-field-xxx
 		$repeaterID = $this->wire('fields')->get('blog_links')->id;
 		$repeaterPage = $pages->get("parent.name=repeaters, name=for-field-$repeaterID");// making sure we are getting the right page
-		if($repeaterPage->id) $pages->delete($repeaterPage, true);
+
+		if($repeaterPage->id) {
+			if($repeaterPage->hasStatus(Page::statusSystem)) $repeaterPage->removeStatus(Page::statusSystem);
+			#$pages->delete($repeaterPage, true);
+		}
 		return $this->cleanUpTemplates();
 	}
 
@@ -142,25 +149,26 @@ class BlogCleanup extends ProcessBlog {
 		$templates = $this->wire('templates');
 
 		$templatesArray = array(
-							'blog',
-							'blog-archives',
-							'blog-authors',
-							'blog-categories',
-							'blog-category',
-							'blog-comments',
-							'blog-links',
-							'blog-post',
-							'blog-posts',
-							'blog-recent-comments',
-							'blog-recent-posts',
-							'blog-recent-tweets',
-							'blog-tag',
-							'blog-tags',
-							'blog-widgets',
-							'blog-widget-basic',
-							'blog-settings',
-							'blog-basic',
-							'repeater_blog-links',
+			'blog',
+			'blog-archives',
+			'blog-authors',
+			'blog-categories',
+			'blog-category',
+			'blog-comments',
+			'blog-links',
+			'blog-post',
+			'blog-posts',
+			'blog-recent-comments',
+			'blog-recent-posts',
+			'blog-recent-tweets',
+			'blog-tag',
+			'blog-tags',
+			'blog-widgets',
+			'blog-widget-basic',
+			'blog-settings',
+			'blog-basic',
+			'repeater_blog-links',
+			'repeater_blog_links',// @note: it seems in PW3, above 'repeater_blog-links' is converted to this, 'repeater_blog_links'
 
 		);
 
@@ -168,25 +176,24 @@ class BlogCleanup extends ProcessBlog {
 		if ($this->blogStyle == 3 || $this->blogStyle == 4)	unset($templatesArray[0]);// blog template
 		if ($this->blogStyle == 2 || $this->blogStyle == 4)	unset($templatesArray[8]);// blog-posts template
 		if ($this->commentsUse !=1) {
-					unset($templatesArray[5]);// blog-comments
-					unset($templatesArray[9]);// blog-recent-comments
-					unset($templatesArray[17]);// blog-basic
+			unset($templatesArray[5]);// blog-comments
+			unset($templatesArray[9]);// blog-recent-comments
+			unset($templatesArray[17]);// blog-basic
 		}
 
 		// delete each found template one by one
 		foreach ($templatesArray as $tpl) {
-					$t = $templates->get($tpl);
-					if ($t->id) {
-							// two step process to delete system template (repeater_blog-links)
-							if ($t->flags == 8) {
-									$t->flags = Template::flagSystemOverride;
-									$t->flags = 0;
-									$templates->delete($t);
-							}
-
-							$templates->delete($t);
-							$this->wire('fieldgroups')->delete($t->fieldgroup);// delete the associated fieldgroups
-					}
+			$t = $templates->get($tpl);
+			if ($t && $t->id) {
+				// two step process to delete system template (repeater_blog-links)
+				if ($t->flags == 8) {
+					$t->flags = Template::flagSystemOverride;
+					$t->flags = 0;
+					//$templates->delete($t);
+				}
+				$templates->delete($t);
+				$this->wire('fieldgroups')->delete($t->fieldgroup);// delete the associated fieldgroups
+			}
 		}
 
 		return $this->cleanUpFields();
@@ -216,36 +223,35 @@ class BlogCleanup extends ProcessBlog {
 
 		// array of blog fields. We'll use this to delete each, one by one as applicable
 		$fieldsArray = array(
-
-						'blog_body',
-						'blog_categories',
-						'blog_comments',
-						'blog_comments_view',
-						'blog_comments_max',
-						'blog_quantity',
-						'blog_date',
-						'blog_files',
-						'blog_headline',
-						'blog_href',
-						'blog_images',
-						'blog_links',
-						'blog_note',
-						'blog_summary',
-						'blog_tags',
-						'blog_small',
+			'blog_body',
+			'blog_categories',
+			'blog_comments',
+			'blog_comments_view',
+			'blog_comments_max',
+			'blog_quantity',
+			'blog_date',
+			'blog_files',
+			'blog_headline',
+			'blog_href',
+			'blog_images',
+			'blog_links',
+			'blog_note',
+			'blog_summary',
+			'blog_tags',
+			'blog_small',
 		);
 
 		// unset irrelevant fields depending on commentsUse (important since user could have a field with similar name that is not part of Blog!)
 		if ($this->commentsUse !=1) {
-					unset($fieldsArray[2]);// blog_comments
-					unset($fieldsArray[3]);// blog_comments_view
-					unset($fieldsArray[4]);// blog_comments_max
+			unset($fieldsArray[2]);// blog_comments
+			unset($fieldsArray[3]);// blog_comments_view
+			unset($fieldsArray[4]);// blog_comments_max
 		}
 
 		// delete each found field
 		foreach ($fieldsArray as $fld) {
-					$f = $fields->get($fld);
-					if($f->id) $fields->delete($f);
+			$f = $fields->get($fld);
+			if($f->id) $fields->delete($f);
 		}
 
 		return $this->cleanUpRoles();
@@ -282,23 +288,23 @@ class BlogCleanup extends ProcessBlog {
 				$this->deleteFiles = true;
 
 				$templateFiles = array(
-										// template files
-										'blog.php',
-										'blog-archives.php',
-										'blog-authors.php',
-										'blog-categories.php',
-										'blog-category.php',
-										'blog-comments.php',
-										'blog-links.php',
-										'blog-post.php',
-										'blog-posts.php',
-										'blog-recent-comments.php',
-										'blog-recent-posts.php',
-										'blog-recent-tweets.php',
-										'blog-side-bar.inc',// will only be present if templateFilesInstall == 2 {demo template files}
-										'blog-tag.php',
-										'blog-tags.php',
-										'blog-main.inc',// will only be present if templateFilesInstall == 2 {demo template files}
+					// template files
+					'blog.php',
+					'blog-archives.php',
+					'blog-authors.php',
+					'blog-categories.php',
+					'blog-category.php',
+					'blog-comments.php',
+					'blog-links.php',
+					'blog-post.php',
+					'blog-posts.php',
+					'blog-recent-comments.php',
+					'blog-recent-posts.php',
+					'blog-recent-tweets.php',
+					'blog-side-bar.inc',// will only be present if templateFilesInstall == 2 {demo template files}
+					'blog-tag.php',
+					'blog-tags.php',
+					'blog-main.inc',// will only be present if templateFilesInstall == 2 {demo template files}
 
 				);
 
@@ -307,19 +313,19 @@ class BlogCleanup extends ProcessBlog {
 				if ($this->blogStyle == 2 || $this->blogStyle == 4) unset($templateFiles[8]);// blog-posts.php
 				if ($this->blogStyle == 3 || $this->blogStyle == 4) unset($templateFiles[0]);// blog.php
 				if ($this->templateFilesInstall !=2) {
-						unset($templateFiles[12]);// blog-side-bar.inc
-						unset($templateFiles[15]);// blog-main.inc
+					unset($templateFiles[12]);// blog-side-bar.inc
+					unset($templateFiles[15]);// blog-main.inc
 				}
 
 				if ($this->commentsUse !=1) {
-							unset($templateFiles[5]);// blog-comments.php
-							unset($templateFiles[9]);// blog-recent-comments.php
+					unset($templateFiles[5]);// blog-comments.php
+					unset($templateFiles[9]);// blog-recent-comments.php
 				}
 
 				// 1. delete template files
 				$sourcepath = $config->paths->templates;// source: '/site/templates/'
 				foreach ($templateFiles as $templateFile) {
-						if(is_file($sourcepath . $templateFile)) unlink($sourcepath . $templateFile);// delete the file if found
+					if(is_file($sourcepath . $templateFile)) unlink($sourcepath . $templateFile);// delete the file if found
 				}
 
 				// 2. delete demo JS file
@@ -361,7 +367,10 @@ class BlogCleanup extends ProcessBlog {
 		$reset = parent::configDefaults();
 
 		// get ProcessBlog class
-		$pb = $modules->get(get_parent_class($this));
+		#$pb = $modules->get(get_parent_class($this));
+		// @note: using strrchr to account for namespaced classes
+		$baseClass = substr(strrchr('\\'.get_parent_class($this), '\\'), 1);
+		$pb = $modules->get($baseClass);
 
 		// save to ProcessBlog config data (reset)
 		$modules->saveModuleConfigData($pb, $reset);
